% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinkREDCap.R
\name{drinkREDCap}
\alias{drinkREDCap}
\alias{loadFromRedcap}
\title{Load data requested into current environment from RedCap}
\usage{
drinkREDCap(
  variables,
  envir = NULL,
  keyring = NULL,
  forms = NULL,
  FUN = sipRedCap,
  config = "auto",
  ...
)

loadFromRedcap(
  variables,
  envir = NULL,
  keyring = NULL,
  forms = NULL,
  FUN = sipRedCap,
  config = "auto",
  ...
)
}
\arguments{
\item{variables}{character vector. A list of strings that define the variables to fill with RedCap data}

\item{envir}{environment. The target environment for the data. Defaults to .Global}

\item{keyring}{character. Potential keyring, not used by default.}

\item{forms}{list. A list of forms. Keys are the variable(api_key), each key can contain a vector of forms.
The output variable is now the <variable>.<form>}

\item{FUN}{function. the function to call. It must have a key argument. If forms are used it should have a forms argument as well.
The default is to call readRC which is a shim for \code{\link[redcapAPI]{exportRecords}}.}

\item{config}{string. Defaults to 'auto'. If set to NULL no configuration file is searched for. If set to anything
but 'auto', that will be the config file override that is used if it exists instead of
searching for the ../<basename>.yml.}

\item{\dots}{Additional arguments passed to FUN.}
}
\value{
Nothing
}
\description{
The first thing it does is check for a yaml config file of
the same name as the current directory with a .yml extension
one level above. This is intended for production environments
where the API_KEY must be stored in a file. If this yaml exists, then it expects this file
to contain `apiUrl` and `apiKeys`. `apiUrl` should be a
string with the URL of the Redcap instance. `apiKeys` should
be a list of variable name keys with values that are their
actual RedCap API_KEY.
}
\details{
Next it will use an api environment in memory to keep api_keys.
If one is knitting with parameters, it will request and store these
keys in memory. Otherwise it will request the user enter
each key using getPass and store it in memory.

IMPORTANT: Make sure that R is set to NEVER save workspace to .RData
as this is the equivalent of writing the API_KEY to a local
file in clear text.
}
\examples{
\donttest{readRC(keyring::get_key()}

}
