% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportMeta.R
\name{exportMeta}
\alias{exportMeta}
\title{exportMeta from REDCap}
\usage{
exportMeta(databases, url, ...)
}
\arguments{
\item{databases}{character; Vector of database names to export Meta Data from.}

\item{url}{character; The url of the REDCap server}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{exportMetaData}}.}
}
\value{
invisible NULL
}
\description{
exportMetaData from REDCap directly into environment variables requested using
API_KEY stored in keyring. Will use the database name provided as the
variable name or the corresponding name if a named vector.
}
\examples{
\dontrun{exportRecords(c("meta"="testDatabase"), keyring="test", url="http://someurl.here")}

}
