% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRecords.R
\name{exportRecords}
\alias{exportRecords}
\title{exportRecords from REDCap}
\usage{
exportRecords(variables = databases, url, ...)
}
\arguments{
\item{url}{character; url of REDCap server}

\item{...}{Additional arguments passed to \code{\link[rccola]{drinkREDCap}}
which in turn passes most to \code{\link[redcapAPI]{exportRecords}}.}

\item{databases}{character; vector of REDCap databases to export records from.
Can be a named vector, and the name is used as the environment variable
to store resulting data in.}
}
\value{
invisible NULL
}
\description{
exportRecords from multiple REDCap databases directly into environment
variables requested using API_KEY stored in keyring.
}
\examples{
\dontrun{exportRecords(c("test"="testDatabase"), keyring="test", url="http://someurl.here")}

}
