% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importRecords.R
\name{importRecords}
\alias{importRecords}
\title{importRecords to REDCap}
\usage{
importRecords(database, data, ...)
}
\arguments{
\item{database}{character; The name given to the REDCap database.}

\item{data}{data.frame; The data.frame of records to write the database.}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{importRecords}}. Should contain url as an argument.}

\item{url}{character; The url of the REDCap server}
}
\value{
invisible NULL
}
\description{
importRecords to REDCap directly into environment variables requested using
API_KEY stored in keyring.
}
\examples{
\dontrun{exportRecords(c("test"="testDatabase"), keyring="test", url="http://someurl.here")}

}
