% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromRedcap.R
\name{loadFromRedcap}
\alias{loadFromRedcap}
\title{Load data requested into current environment from RedCap}
\usage{
loadFromRedcap(variables, apiUrl = "https://redcap.vanderbilt.edu/api/")
}
\arguments{
\item{variables}{A list of strings that define the variables to fill with RedCap data}

\item{apiUrl}{The api interface to the RedCap instance to use. defaults to the Vanderbilt instance.}
}
\value{
Nothing
}
\description{
The first thing it does is check for a yaml config file of
the same name as the current directory with a .yml extension
one level above. This is intended for production environments
where the API_KEY must be stored in a file. If this yaml exists, then it expects this file
to contain `apiUrl` and `apiKeys`. `apiUrl` should be a
string with the URL of the Redcap instance. `apiKeys` should
be a list of variable name keys with values that are their
actual RedCap API_KEY.
}
\details{
Next it will use an api environment in memory to keep api_keys.
If one is knitting with parameters, it will request and store these
keys in memory. Otherwise it will request the user enter
each key using getPass and store it in memory.

IMPORTANT: Make sure that R is set to NEVER save workspace to .RData
as this is the equivalent of writing the API_KEY to a local
file in clear text.
}
\examples{
\donttest{loadFromRedcap(data)}

}
