% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinkREDCap.R
\name{metaREDCap}
\alias{metaREDCap}
\title{Alternate metadata function to read just the meta data from REDCap}
\usage{
metaREDCap(key, ...)
}
\arguments{
\item{key}{the api key of interest. The package provides this.}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{exportMetaData}}. Should contain url as an argument.}
}
\value{
data.frame containing requested REDCap data.
}
\description{
Alternate metadata function to read just the meta data from REDCap
}
\examples{
\dontrun{data <- metaREDCap(keyring::key_get("rccola", "database_name", "project_name"))}

}
