% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCdeleteRecords.R
\name{RCdeleteRecords}
\alias{RCdeleteRecords}
\title{Import records to REDCap}
\usage{
RCdeleteRecords(database, records, keyring, url, ...)
}
\arguments{
\item{database}{character; The name given to the REDCap database.}

\item{records}{character; A vector of record_ids to delete}

\item{keyring}{character; Name of keyring}

\item{url}{character; The url of the REDCap server}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{deleteRecords}}. Should contain url as an argument.}
}
\value{
invisible NULL
}
\description{
Impmort records to REDCap directly and return results in environment variables
requested using API_KEY stored in keyring.
}
\examples{
\dontrun{RCdeleteRecords(c("test"="testDatabase"), keyring="test", url="http://someurl.here")}

}
