% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCexportMeta.R
\name{RCexportMeta}
\alias{RCexportMeta}
\title{Export metadata from REDCap}
\usage{
RCexportMeta(databases, keyring, url, ...)
}
\arguments{
\item{databases}{character; Vector of database names to export Meta Data from.}

\item{keyring}{character; The keyring to access to get API_KEY associated with database name.}

\item{url}{character; The url of the REDCap server}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{exportMetaData}}.}
}
\value{
invisible NULL
}
\description{
Export metadata from REDCap directly into environment variables requested using
API_KEY stored in keyring. Will use the database name provided as the
variable name or the corresponding name if a named vector.
}
\examples{
\dontrun{RCexportMeta(c("meta"="testDatabase"), keyring="test", url="http://someurl.here")}

}
