% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCexportRecords.R
\name{RCexportRecords}
\alias{RCexportRecords}
\title{Export records from REDCap}
\usage{
RCexportRecords(databases, keyring, url, ...)
}
\arguments{
\item{databases}{character; vector of REDCap databases to export records from.
Can be a named vector, and the name is used as the environment variable
to store resulting data in.}

\item{keyring}{character; keyring to use to find API_KEY of database}

\item{url}{character; url of REDCap server}

\item{...}{Additional arguments passed to \code{\link[rccola]{drinkREDCap}}
which in turn passes most to \code{\link[redcapAPI]{exportRecords}}.}
}
\value{
invisible NULL
}
\description{
Export records from multiple REDCap databases directly into environment
variables requested using API_KEY stored in keyring.
}
\examples{
\dontrun{RCexportRecords(c("test"="testDatabase"), keyring="test", url="http://someurl.here")}

}
