% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCimportRecords.R
\name{RCimportRecords}
\alias{RCimportRecords}
\title{Import records to REDCap}
\usage{
RCimportRecords(database, data, keyring, url, ...)
}
\arguments{
\item{database}{character; The name given to the REDCap database.}

\item{data}{data.frame; The data.frame of records to write the database.}

\item{keyring}{character; Keyring to look up database API_KEY from}

\item{url}{character; The url of the REDCap server}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{importRecords}}. Should contain url as an argument.}
}
\value{
invisible NULL
}
\description{
Import records to REDCap directly into environment variables requested using
API_KEY stored in keyring.
}
\examples{
\dontrun{exportRecords(c("test"="testDatabase"), keyring="test", url="http://someurl.here")}

}
