% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFromRedcap.R
\name{readRC}
\alias{readRC}
\title{Default function to read from redcap}
\usage{
readRC(key, ...)
}
\arguments{
\item{key}{the api key of interest. The package provides this.}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{exportRecords}}.}
}
\description{
Default function to read from redcap
}
\examples{
\donttest{data <- readRC(keyring::key_get("rccola", "database_name", "project_name"))}

}
