% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drinkREDCap.R
\name{sipREDCap}
\alias{sipREDCap}
\title{Default function to read from REDCap}
\usage{
sipREDCap(key, ...)
}
\arguments{
\item{key}{the api key of interest. The package provides this.}

\item{...}{Additional arguments passed to \code{\link[redcapAPI]{exportRecords}}.}
}
\value{
data.frame containing requested REDCap data.
}
\description{
Default function to read from REDCap
}
\examples{
\dontrun{data <- sipREDCap(keyring::key_get("rccola", "database_name", "project_name"))}

}
